The doctype _must_ be defined to avoid 'quirks mode'
This is an issue that plagued me this weekend. For flex grand-children of a flex-item whose main parent had align-items:stretch, the height being 100% continuously was being set to 100% of the entire viewport. Why is this? It turns out to be because without setting the DOCTYPE, browsers will revert to "quirks mode" which was aimed at being compatible with IE5 and lower.